Only refresh the owd_test_cases if the caller sets a variable saying so (./install.sh).

+++ install.sh
@@ -49,6 +49,8 @@ fi
 #
 [ ! -d ./gaia-ui-tests ] && ./bin/setup_gaiatest.sh "$BRANCH"
 
+
+
 #
 # Install me.
 #
@@ -71,17 +73,25 @@ cd $HOME/projects/OWD_TEST_TOOLKIT
 sudo python setup.py install >> $LOGFILE
 
 
-#
-# Install the owd test cases.
-#
-printf "\n\nInstalling owd_test_cases..." | tee -a $LOGFILE
-printf "\n============================\n" | tee -a $LOGFILE
-cd $HOME/projects
-rm -rf owd_test_cases 2>/dev/null
 
-git clone https://github.com/roydude/owd_test_cases.git >> $LOGFILE 2>> $LOGFILE
-cd $HOME/projects/owd_test_cases
+#
+# If you're developing the toolkit, this might be the last thing you want,
+# so I've defaulted it to not do this! ...
+#
+if [ "$REINSTALL_OWD_TEST_CASES" ]
+then
+    #
+    # Install the owd test cases.
+    #
+    printf "\n\nInstalling owd_test_cases..." | tee -a $LOGFILE
+    printf "\n============================\n" | tee -a $LOGFILE
+    cd $HOME/projects
+    rm -rf owd_test_cases 2>/dev/null
 
-printf "\n* Switching to branch $BRANCH of owd_test_cases ...\n\n" | tee -a $LOGFILE
-git checkout $BRANCH  2> >( tee -a $LOGFILE)
-printf "\n* Now using owd_test_cases branch \"$(git branch | grep '*')\".\n\n"
+    git clone https://github.com/roydude/owd_test_cases.git >> $LOGFILE 2>> $LOGFILE
+    cd $HOME/projects/owd_test_cases
+
+    printf "\n* Switching to branch $BRANCH of owd_test_cases ...\n\n" | tee -a $LOGFILE
+    git checkout $BRANCH  2> >( tee -a $LOGFILE)
+    printf "\n* Now using owd_test_cases branch \"$(git branch | grep '*')\".\n\n"
+fi
